+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Add support for non-python text files
+- Add a built-in corpus of software-related terms
+- `-v` flag to turn on printing as files are checked
+- New flag to show words not in the corpus but also not identifiable as typos
+- New flag to load system dict on linux boxes with dict in `/usr/share/`
+
 ## 0.0.1
 
 - Initial version. Python support only.